CV and Contact pages added and routed
