+++ .firebase/hosting.YnVpbGQ.cache
@@ -1,32 +1,32 @@
-asset-manifest.json,1598986842525,2118821906fa223d33acda96313a061259d7981a42a9a5ae2f01f1d173177d0b
-index.html,1598986842524,7145cb5c7fea730f7ab21d1b60dbb30d67f00806b7036f709235c07b021d8602
-precache-manifest.f508403d1cd8f0fabac1643ccb14c095.js,1598986842524,3983d363f36cdd54ae648e8c550828574de78a84ad172906ab30bb6c852218b3
-favicon.ico,1598986835643,a08fa4488c3ecef62d9effd03b3a989929bdcbecf5e905941f9034a15bd3dba3
-logo192.png,1598986835714,caff018b7f1e8fd481eb1c50d75b0ef236bcd5078b1d15c8bb348453fee30293
-service-worker.js,1598986842525,37949a40a47870e85552d64e8bee571bafbd461fbf82f0d544c254e5ed1b5379
-manifest.json,1598986835716,341d52628782f8ac9290bbfc43298afccb47b7cbfcee146ae30cf0f46bc30900
-robots.txt,1598986835717,391d14b3c2f8c9143a27a28c7399585142228d4d1bdbe2c87ac946de411fa9a2
-logo512.png,1598986835715,191fc21360b4ccfb1cda11a1efb97f489ed22672ca83f4064316802bbfdd750e
-static/css/main.5f361e03.chunk.css,1598986842525,26f63cbd305dee0938a571297979412133c617131502dd537b087a8f85fe24ec
-static/css/main.5f361e03.chunk.css.map,1598986842528,c4765569ad99b9b74db97a058a3c9b5068c9f8d7904b4a63857f5a65a560abed
-static/js/2.d8a0f125.chunk.js.LICENSE.txt,1598986842529,5510346277836fc9811b2eee7ff015390cdf98f869f66b87219955660acecbc7
-static/js/main.8080691c.chunk.js,1598986842528,0add88909714d2d16acf51ef6c70ecdb9722a7b6c82ff4debff6ff1e74c1cc0a
-static/js/runtime-main.6a16053d.js,1598986842528,73c59721133cf88768cf53033ad53fc2713cb897d7bc54a73536550ab277785c
-static/js/main.8080691c.chunk.js.map,1598986842529,14b8e8a7102e69d73947e58691efc4f9d16986b48acd639c2a077aee27657149
-static/js/runtime-main.6a16053d.js.map,1598986842530,a1a8178c956ed503808cec6fc19da8f0005a26c672be4ae87aebbc7e3061c39e
-imgs/MLAirHockey/AirHockeyModel.jpeg,1598986835650,67c3b45a93ae4adf59415b59a34f183edaa469fabdd39a9848f0b9a9e88f31f3
-imgs/MLAirHockey/MLSim.png,1598986835651,a7d2d85f21489d056ea7990cf2dc5f2230944c485d76ac0287366fbb55dd2396
-imgs/MLAirHockey/hitrate.png,1598986835661,3fe23ca056917cc0ee702fe5b3db99b18eed8cd8585bbb6effe6adcd0ccad8b7
-static/css/2.11829350.chunk.css,1598986842528,60f20ac67d438a4a83cabd219ef5f8a48e8594011f67be2a8ddc92cc604ff02a
-static/js/2.d8a0f125.chunk.js,1598986842529,502ddf9e616b541e6b16aaaa4d280b2991e4091119a6596c098b054b36701652
-imgs/2TapTimer/timerSelect.PNG,1598986835648,aee717e56d82053bdae03a7d5670a1ba144c0c8b048509ec147ab03f09562acd
-imgs/PortfolioSite/portfolioSite.png,1598986835710,7cc40515abd089f3c004ed26aecc0c0e432060a143592ae68f43ff06c98272b5
-imgs/MLAirHockey/stick_model.JPG,1598986835676,fca33de3f5d93528e7d8fda1521d976f334de0fa820e050418afe6432a70edfb
-static/css/2.11829350.chunk.css.map,1598986842529,44addd0ee7cc243f4451072b74a0ddf37b7a93c18ab123608ff67d074f861db3
-imgs/PillowFluffer/Overhead.png,1598986835693,128ff3911eb1120e50b2e6165b98a3871badb46d3efe3d592f8cabf157478319
-imgs/PillowFluffer/pov.png,1598986835699,4179d53d3f2b712a7995134a9c5069291368ca53a3c950f7edd130a91dbf914d
-static/js/2.d8a0f125.chunk.js.map,1598986842529,588fdcdc375b48195b0761c8ed99d8c8824c6c59190b01d0afba7378992b0a91
-imgs/PillowFluffer/unity.png,1598986835707,bbdf10630bd88ae6e218164090d0ee8f3ef4836ab8916fa64e388de968bf1ea5
-imgs/MLAirHockey/Poster.png,1598986835659,120a84531875bf99343067e981f9346de7d41a03fa03da27c29606d7d474ca1c
-imgs/MLAirHockey/stick.jpg,1598986835670,d5e2afe267554b19ec781a444362703fdefaa1a549eed86c573f3d6974e9d091
-imgs/MLAirHockey/table.jpg,1598986835686,f47fb2fa30d57399d4548dd455a17488d4dd8d1867aebc45019c8286b3b42289
+asset-manifest.json,1599002374669,2118821906fa223d33acda96313a061259d7981a42a9a5ae2f01f1d173177d0b
+precache-manifest.f508403d1cd8f0fabac1643ccb14c095.js,1599002374669,3983d363f36cdd54ae648e8c550828574de78a84ad172906ab30bb6c852218b3
+favicon.ico,1599002368346,a08fa4488c3ecef62d9effd03b3a989929bdcbecf5e905941f9034a15bd3dba3
+service-worker.js,1599002374669,37949a40a47870e85552d64e8bee571bafbd461fbf82f0d544c254e5ed1b5379
+logo192.png,1599002368385,caff018b7f1e8fd481eb1c50d75b0ef236bcd5078b1d15c8bb348453fee30293
+index.html,1599002374669,7145cb5c7fea730f7ab21d1b60dbb30d67f00806b7036f709235c07b021d8602
+robots.txt,1599002368387,391d14b3c2f8c9143a27a28c7399585142228d4d1bdbe2c87ac946de411fa9a2
+logo512.png,1599002368386,191fc21360b4ccfb1cda11a1efb97f489ed22672ca83f4064316802bbfdd750e
+manifest.json,1599002368386,341d52628782f8ac9290bbfc43298afccb47b7cbfcee146ae30cf0f46bc30900
+static/css/main.5f361e03.chunk.css,1599002374670,26f63cbd305dee0938a571297979412133c617131502dd537b087a8f85fe24ec
+static/css/main.5f361e03.chunk.css.map,1599002374675,c4765569ad99b9b74db97a058a3c9b5068c9f8d7904b4a63857f5a65a560abed
+static/js/2.d8a0f125.chunk.js.LICENSE.txt,1599002374675,5510346277836fc9811b2eee7ff015390cdf98f869f66b87219955660acecbc7
+static/js/main.8080691c.chunk.js,1599002374674,0add88909714d2d16acf51ef6c70ecdb9722a7b6c82ff4debff6ff1e74c1cc0a
+static/js/runtime-main.6a16053d.js,1599002374675,73c59721133cf88768cf53033ad53fc2713cb897d7bc54a73536550ab277785c
+static/js/main.8080691c.chunk.js.map,1599002374675,14b8e8a7102e69d73947e58691efc4f9d16986b48acd639c2a077aee27657149
+static/js/runtime-main.6a16053d.js.map,1599002374675,a1a8178c956ed503808cec6fc19da8f0005a26c672be4ae87aebbc7e3061c39e
+imgs/MLAirHockey/AirHockeyModel.jpeg,1599002368350,67c3b45a93ae4adf59415b59a34f183edaa469fabdd39a9848f0b9a9e88f31f3
+imgs/MLAirHockey/MLSim.png,1599002368352,a7d2d85f21489d056ea7990cf2dc5f2230944c485d76ac0287366fbb55dd2396
+imgs/MLAirHockey/hitrate.png,1599002368357,3fe23ca056917cc0ee702fe5b3db99b18eed8cd8585bbb6effe6adcd0ccad8b7
+static/css/2.11829350.chunk.css,1599002374674,60f20ac67d438a4a83cabd219ef5f8a48e8594011f67be2a8ddc92cc604ff02a
+static/js/2.d8a0f125.chunk.js,1599002374675,502ddf9e616b541e6b16aaaa4d280b2991e4091119a6596c098b054b36701652
+imgs/2TapTimer/timerSelect.PNG,1599002368349,aee717e56d82053bdae03a7d5670a1ba144c0c8b048509ec147ab03f09562acd
+imgs/PortfolioSite/portfolioSite.png,1599002368384,7cc40515abd089f3c004ed26aecc0c0e432060a143592ae68f43ff06c98272b5
+imgs/MLAirHockey/stick_model.JPG,1599002368365,fca33de3f5d93528e7d8fda1521d976f334de0fa820e050418afe6432a70edfb
+static/css/2.11829350.chunk.css.map,1599002374675,44addd0ee7cc243f4451072b74a0ddf37b7a93c18ab123608ff67d074f861db3
+imgs/PillowFluffer/Overhead.png,1599002368374,128ff3911eb1120e50b2e6165b98a3871badb46d3efe3d592f8cabf157478319
+imgs/PillowFluffer/pov.png,1599002368377,4179d53d3f2b712a7995134a9c5069291368ca53a3c950f7edd130a91dbf914d
+imgs/PillowFluffer/unity.png,1599002368382,bbdf10630bd88ae6e218164090d0ee8f3ef4836ab8916fa64e388de968bf1ea5
+static/js/2.d8a0f125.chunk.js.map,1599002374675,588fdcdc375b48195b0761c8ed99d8c8824c6c59190b01d0afba7378992b0a91
+imgs/MLAirHockey/Poster.png,1599002368355,120a84531875bf99343067e981f9346de7d41a03fa03da27c29606d7d474ca1c
+imgs/MLAirHockey/stick.jpg,1599002368361,d5e2afe267554b19ec781a444362703fdefaa1a549eed86c573f3d6974e9d091
+imgs/MLAirHockey/table.jpg,1599002368370,f47fb2fa30d57399d4548dd455a17488d4dd8d1867aebc45019c8286b3b42289
