+++ src/pages/CV.tsx
@@ -1,0 +1,20 @@
+import React, {useState} from 'react';
+import { Button, Alert } from 'react-bootstrap';
+
+
+// import poster from '../imgs/MLAirHockey/Poster.png'
+
+interface CVProps {
+
+};
+
+let x:number = 0;
+const CV: React.FC<CVProps> = (CVProps) => {
+    return(
+        <>
+            <iframe src="https://docs.google.com/document/d/1wF3zutNxZVcCZicvzA6lWvpWTbLontrPdGXB7OkF3y8/edit?usp=sharing" frameBorder="1" style={{ display:"block", width:"100%", height:"1000px"}}>IFRAME not supported!</iframe>
+        </>
+    );
+}
+
+export default CV
