+++ src/siteImages.json
@@ -1,22 +1,22 @@
 {
     "twoTapTimer":[
-        "imgs/2TapTimer/timerSelect.PNG"
+        "../imgs/2TapTimer/timerSelect.PNG"
     ],
     "mlAirHockey":[
-        "imgs/MLAirHockey/AirHockeyModel.jpeg",
-        "imgs/MLAirHockey/hitrate.png",
-        "imgs/MLAirHockey/MLSim.png",
-        "imgs/MLAirHockey/Poster.png",
-        "imgs/MLAirHockey/stick.jpg",
-        "imgs/MLAirHockey/stick_model.JPG",
-        "imgs/MLAirHockey/table.jpg"
+        "../imgs/MLAirHockey/AirHockeyModel.jpeg",
+        "../imgs/MLAirHockey/hitrate.png",
+        "../imgs/MLAirHockey/MLSim.png",
+        "../imgs/MLAirHockey/Poster.png",
+        "../imgs/MLAirHockey/stick.jpg",
+        "../imgs/MLAirHockey/stick_model.JPG",
+        "../imgs/MLAirHockey/table.jpg"
     ],
     "pillowFluffer":[
-        "imgs/PillowFluffer/Overhead.png",
-        "imgs/PillowFluffer/pov.png",
-        "imgs/PillowFluffer/unity.png"
+        "../imgs/PillowFluffer/Overhead.png",
+        "../imgs/PillowFluffer/pov.png",
+        "../imgs/PillowFluffer/unity.png"
     ],    
     "portfolioSite": [
-        "imgs/PortfolioSite/portfolioSite.png"
+        "../imgs/PortfolioSite/portfolioSite.png"
     ]
 }
